:building_construction: Move constants into own file
