+++ src/date.rs
@@ -1,7 +1,7 @@
 use crate::{
     errors::{out_of_range::create_simple_oor, AstrolabeError},
-    shared::{DAYS_TO_1970, DAYS_TO_1970_I64, SECS_PER_DAY_U64},
     util::{
+        constants::{DAYS_TO_1970, DAYS_TO_1970_I64, SECS_PER_DAY_U64},
         date::{
             convert::{date_to_days, days_to_date, year_doy_to_days},
             manipulate::{apply_date_unit, set_date_unit},
