+++ src/util/format.rs
@@ -1,9 +1,8 @@
-use crate::shared::{
-    MONTH_ABBREVIATED, MONTH_NARROW, MONTH_WIDE, NANOS_PER_SEC, SECS_PER_DAY, SECS_PER_HOUR,
-    SECS_PER_MINUTE, WDAY_ABBREVIATED, WDAY_NARROW, WDAY_SHORT, WDAY_WIDE,
-};
-
 use super::{
+    constants::{
+        MONTH_ABBREVIATED, MONTH_NARROW, MONTH_WIDE, NANOS_PER_SEC, SECS_PER_DAY, SECS_PER_HOUR,
+        SECS_PER_MINUTE, WDAY_ABBREVIATED, WDAY_NARROW, WDAY_SHORT, WDAY_WIDE,
+    },
     date::convert::{days_to_date, days_to_doy, days_to_wday, days_to_wyear},
     time::convert::nanos_to_time,
 };
